Allows processing of directives after init
Previously the document was scanned for directives at the beginning and
beyond that, only redrawing of the page was done. This is not practical
as elements containing directives can be created on the fly - repeating
elements are a good example.

The directive scanning process has been moved in to the templater and a
single call to render() will now both look for new directives and redraw
anything within the {{ brace }} syntax. The initialiser class has been
removed.